+++ packages/shared/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@grassroots/shared",
+  "version": "1.0.0",
+  "description": "Shared types and utilities for Grassroots application",
+  "main": "src/shared.ts",
+  "types": "src/shared.ts",
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch",
+    "clean": "rm -rf dist"
+  },
+  "dependencies": {
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.1"
+  },
+  "devDependencies": {
+    "typescript": "^5.7.3"
+  }
+}
