+++ packages/shared/src/LoginState.dto.ts
