+++ packages/shared/src/Void.dto.ts
