+++ packages/shared/src/shared.ts
@@ -1,0 +1,21 @@
+// DTOs
+export * from "./Contact.dto";
+export * from "./Hello.dto";
+export * from "./LoginState.dto";
+export * from "./Organization.dto";
+export * from "./Paginated.dto";
+export * from "./Permission.dto";
+export * from "./Role.dto";
+export * from "./User.dto";
+export * from "./Void.dto";
+
+// Utilities
+export * from "./util/AssertPropsEqual";
+export * from "./util/Cast";
+export * from "./util/CreateDTOBase";
+export * from "./util/NotNull";
+export * from "./util/PropsOf";
+
+// Other exports
+export * from "./Tree";
+export * from "./OpenAPI.gen";
