+++ packages/shared/src/util/Cast.ts
