+++ packages/shared/src/util/Fail.ts
