+++ packages/shared/src/util/NotNull.ts
