+++ packages/shared/src/util/PropsOf.ts
